+++ src/amd/magix.js
@@ -92,6 +92,9 @@ define('magix', ['$'], $ => {
     /*#if(modules.router||modules.state){#*/
     Inc('../tmpl/dispatcher');
     /*#}#*/
+    /*#if(modules.updater&&modules.updaterAsync){#*/
+    Inc('../tmpl/async');
+    /*#}#*/
     Inc('../tmpl/vframe');
     /*#if(modules.nodeAttachVframe){#*/
     $.fn.invokeView = function (name, args) {
@@ -114,13 +117,10 @@ define('magix', ['$'], $ => {
     /*#}#*/
     Inc('../tmpl/body');
     /*#if(modules.updater){#*/
-    /*#if(!modules.updaterVDOM&&!modules.updaterDOM&&!modules.updaterVRDOM){#*/
+    /*#if(!modules.updaterVDOM&&!modules.updaterDOM){#*/
     Inc('../tmpl/tmpl');
     /*#}#*/
-    /*#if(modules.updaterVRDOM){#*/
-    Inc('../tmpl/tovdom');
-    Inc('../tmpl/vrdom');
-    /*#}else if(modules.updaterVDOM){#*/
+    /*#if(modules.updaterVDOM){#*/
     Inc('../tmpl/tovdom');
     Inc('../tmpl/vdom');
     /*#}else if(modules.updaterDOM){#*/