+++ src/kissy/magix.js
@@ -60,6 +60,9 @@ KISSY.add('magix', (S, SE, DOM) => {
     /*#if(modules.router||modules.state){#*/
     Inc('../tmpl/dispatcher');
     /*#}#*/
+    /*#if(modules.updater&&modules.updaterAsync){#*/
+    Inc('../tmpl/async');
+    /*#}#*/
     Inc('../tmpl/vframe');
     /*#if(modules.nodeAttachVframe){#*/
     DOM[G_PROTOTYPE].invokeView = function (name, args) {
@@ -84,13 +87,10 @@ KISSY.add('magix', (S, SE, DOM) => {
     Inc('../tmpl/body');
 
     /*#if(modules.updater){#*/
-    /*#if(!modules.updaterVDOM&&!modules.updaterDOM&&!modules.updaterVRDOM){#*/
+    /*#if(!modules.updaterVDOM&&!modules.updaterDOM){#*/
     Inc('../tmpl/tmpl');
     /*#}#*/
-    /*#if(modules.updaterVRDOM){#*/
-    Inc('../tmpl/tovdom');
-    Inc('../tmpl/vrdom');
-    /*#}else if(modules.updaterVDOM){#*/
+    /*#if(modules.updaterVDOM){#*/
     Inc('../tmpl/tovdom');
     Inc('../tmpl/vdom');
     /*#}else if(modules.updaterDOM){#*/